+++ Protocols/SecureShuffle.h
@@ -104,7 +104,7 @@ public:
     void apply(vector<T>& a, size_t n, int unit_size, size_t output_base,
             size_t input_base, shuffle_type& shuffle, bool reverse);
 
-    void applyMultiple(vector<T>& a, vector<int>& sizes, vector<int>& destinations, vector<int>& sources,
+    void applyMultiple(vector<T>& a, vector<size_t>& sizes, vector<int>& destinations, vector<int>& sources,
                        vector<int>& unit_sizes, vector<int>& handles, vector<bool>& reverse, store_type& store);
 
     /**
